myProg/proxy_backend/requirements.txt
channels==3.0.4
chardet==3.0.4
Django==3.0.3
func_timeout==4.3.5
paramiko==2.10.3
pyquery==1.4.3
python_dateutil==2.8.2
pytz==2019.3
Requests==2.31.0
